+++ RELEASENOTES.md
@@ -96,6 +96,13 @@
     *   Implement `getDeviceInfo()` to be able to identify `CastPlayer` when
         controlling playback with a `MediaController`
         ([#142](https://github.com/androidx/media/issues/142)).
+*   Remove deprecated symbols:
+    *   Remove `DefaultAudioSink` constructors, use `DefaultAudioSink.Builder`
+        instead.
+*   Remove deprecated symbols:
+    *   Remove `Transformer.Builder.setOutputMimeType(String)`. This feature
+        has been removed. The MIME type will always be MP4 when the default
+        muxer is used.
 
 ### 1.0.0-beta02 (2022-07-21)